Combine grid and "align to closest" command for better interop
This allows using the left and right buttons to adjust after aligning using the grid, and is generally cleaner.
